resep obat

+++ app/Http/Controllers/PasienController.php
@@ -9,11 +9,25 @@ class PasienController extends Controller
 {
     public function index()
     {
-        return Pasien::all();
+        if  (Pasien::all()->isNotEmpty())
+        {
+            return Pasien::all();
+        }
+        else
+        {
+            return 404;
+        }
     }
     public function show($id_pasien)
     {
-        return Pasien::find($id_pasien);
+        if (Pasien::find($id_pasien))
+        {
+            return Pasien::find($id_pasien);
+        }
+        else
+        {
+            return 404;
+        }
     }
     public function store(Request $request)
     {
@@ -22,13 +36,27 @@ class PasienController extends Controller
     public function update(Request $request, $id_pasien)
     {
         $pasien = Pasien::findOrFail($id_pasien);
-        $pasien->update($request->all());
-        return $pasien;
+        if ($pasien)
+        {
+            $pasien->update($request->all());
+            return $pasien;
+        } 
+        else
+        {
+            return 404;
+        }
     }
     public function delete(Request $request, $id_pasien)
     {
         $pasien = Pasien::findOrFail($id_pasien);
-        $pasien->delete();
-        return 200;
+        if ($pasien)
+        {
+            $pasien->delete();
+            return 200;
+        } 
+        else
+        {
+            return 404;
+        }
     }
 }
